change value type to int dont want decimal

+++ core/src/com/mygdx/game/engine/components/LevelComponent.java
@@ -3,5 +3,5 @@ package com.mygdx.game.engine.components;
 import com.badlogic.ashley.core.Component;
 
 public class LevelComponent implements Component {
-    public float level = 1;
+    public int level = 1;
 }
